+++ src/contracts/atlas/DAppVerification.sol
@@ -21,7 +21,7 @@ import "forge-std/Test.sol"; // TODO remove
 // frontend
 contract DAppVerification is EIP712, DAppIntegration {
     using ECDSA for bytes32;
-    using CallBits for uint16;
+    using CallBits for uint32;
 
     bytes32 public constant DAPP_TYPE_HASH = keccak256(
         "DAppProof(address from,address to,uint256 nonce,uint256 deadline,bytes32 userOpHash,bytes32 callChainHash,bytes32 controlCodeHash)"
