+++ test/Permit69.t.sol
@@ -287,7 +287,7 @@ contract MockAtlasForPermit69Tests is Permit69 {
         address user,
         bytes32 controlCodeHash,
         address controller,
-        uint16 callConfig
+        uint32 callConfig
     ) internal view virtual override returns (address activeEnvironment) {
         activeEnvironment = _environment;
     }
